feat(products): add server actions to manipulate data
Create dialog and alert dialog components to update and delete products.
Change react-query keys to be set in one source file.
Create a function to ensure that user is Admin.
Update get-params file to be able to set a default order and returns orderBy date as default.
Some minor changes on styles in toast, skeleton and dialogs.